# src/swagger.json
{
	"components": {
...
		"schemas": {
			"CheckInRes": {
				"properties": {
					"card_no": {
						"type": "number",
						"format": "double",
						"description": "Card number"
					},
					"state": {
						"type": "string",
						"description": "The User state after checkin"
					},
					"prev_state": {
						"type": "string",
						"description": "The Previous user state before checkin. The value is 'checkIn' if duplicate checkin operation."
					},
					"notice": {
						"type": "boolean",
						"description": "A discord notification result when an available count is less than 5"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResJson_CheckInRes_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"payload": {
						"$ref": "#/components/schemas/CheckInRes"
					}
				},
				"required": [
					"status",
					"code",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorJson": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"code",
					"result",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IConfig": {
				"properties": {
					"_id": {
						"type": "number",
						"format": "double"
					},
					"env": {
						"type": "string"
					},
					"auth": {
						"type": "string"
					},
					"begin_at": {
						"type": "string",
						"format": "date-time"
					},
					"end_at": {
						"type": "string",
						"format": "date-time"
					},
					"open_at": {
						"type": "number",
						"format": "double"
					},
					"close_at": {
						"type": "number",
						"format": "double"
					},
					"checkin_at": {
						"type": "number",
						"format": "double"
					},
					"checkout_at": {
						"type": "number",
						"format": "double"
					},
					"seocho": {
						"type": "number",
						"format": "double"
					},
					"gaepo": {
						"type": "number",
						"format": "double"
					},
					"actor": {
						"type": "string"
					},
					"deleted_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResJson_IConfig_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"payload": {
						"$ref": "#/components/schemas/IConfig"
					}
				},
				"required": [
					"status",
					"code",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConfigUpdate": {
				"properties": {
					"gaepo": {
						"type": "number",
						"format": "double"
					},
					"seocho": {
						"type": "number",
						"format": "double"
					},
					"begin_at": {
						"type": "string",
						"format": "date-time"
					},
					"end_at": {
						"type": "string",
						"format": "date-time"
					},
					"open_at": {
						"type": "number",
						"format": "double"
					},
					"close_at": {
						"type": "number",
						"format": "double"
					},
					"auth": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ConfigRequest": {
				"properties": {
					"env": {
						"$ref": "#/components/schemas/ConfigUpdate"
					},
					"values": {
						"$ref": "#/components/schemas/ConfigUpdate"
					},
					"date": {
						"type": "string"
					}
				},
				"required": [
					"date"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"values": {
						"end_at": "2022-02-06",
						"open_at": "08:00:00",
						"close_at": "22:00:00",
						"seocho": 300,
						"gaepo": 400
					},
					"date": "2022-01-25"
				}
			},
...
		},
		"securitySchemes": {
			"CookieAuth": {
				"type": "apiKey",
				"name": "w_auth_local",
				"in": "cookie"
			},
			"OAuth42": {
				"type": "oauth2",
				"flows": {
					"clientCredentials": {
						"type": "oauth2",
						"authorizationUrl": "http://localhost:3000/user/login",
						"tokenUrl": "http://localhost:3000/user/login/token",
						"refreshUrl": "http://localhost:3000/user/token/refresh",
						"scopes": []
					}
				}
			}
		}
	},
	"info": {
		"title": "42checkin",
		"version": "2.2.7",
		"description": "Cluster checkin service for 42 Seoul cadets.",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "@ryukim, @yurlee, @ohjongin"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/user/checkIn/{cardId}": {
			"post": {
				"operationId": "CheckIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResJson_CheckInRes_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 200,
											"result": true,
											"code": 2000,
											"payload": {
												"card_no": 10,
												"state": "checkIn",
												"prev_state": "checkOut",
												"notice": false
											}
										}
									},
									"Example 2": {
										"value": {
											"status": 200,
											"result": false,
											"code": 4090,
											"message": "이미 체크인 상태입니다",
											"payload": {
												"card_no": 10,
												"state": "checkIn",
												"prev_state": "checkIn",
												"notice": false
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 401,
											"code": 4010,
											"message": "사용자 정보 없음",
											"result": false
										}
									}
								}
							}
						}
					},
					"406": {
						"description": "Not Acceptable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 406,
											"code": 4060,
											"message": "체크인 가능 시간이 아닙니다.",
											"result": false
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "Not Acceptable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 409,
											"code": 4090,
											"message": "이미 사용 중인 카드입니다. cardOwner: cadet",
											"result": false
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 500,
											"code": 5000,
											"message": "Internal Server Error",
											"result": false
										}
									}
								}
							}
						}
					}
				},
				"description": "유저 체크인",
				"tags": [
					"User"
				],
				"security": [
					{
						"CookieAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cardId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/v1/config": {
			"get": {
				"operationId": "GetConfig",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResJson_IConfig_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 200,
											"code": 2000,
											"result": true,
											"payload": {
												"_id": 1,
												"env": "production",
												"auth": "42",
												"begin_at": "2021-12-31T15:00:00.000Z",
												"end_at": "2022-02-20T00:00:00.000Z",
												"open_at": "08:00:00",
												"close_at": "22:00:00",
												"checkin_at": "08:00:00",
												"checkout_at": "22:00:00",
												"seocho": 200,
												"gaepo": 300,
												"actor": "chaepark",
												"deleted_at": null,
												"updated_at": "2022-01-25T08:24:28.376Z",
												"created_at": "2021-12-31T15:00:00.000Z",
												"_comment": 1
											}
										}
									},
									"Example 2": {
										"value": {
											"status": 200,
											"code": 4040,
											"result": false,
											"message": "[production] 해당 날짜(2022-01-31)의 설정값이 서버에 존재하지 않습니다."
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 500,
											"code": 5000,
											"result": false,
											"message": "서버 오류입니다.\n잠시 후 재시도하시거나,\n증상이 계속되면 관리자(slack:@ohjongin)에게 문의하세요."
										}
									}
								}
							}
						}
					}
				},
				"description": "클러스터 설정 값 조회",
				"tags": [
					"Config"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "date",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "2022-01-25"
					}
				]
			},
			"put": {
				"operationId": "SetConfig",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResJson_IConfig_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 200,
											"code": 200,
											"result": true,
											"payload": {
												"_id": 1,
												"env": "local",
												"auth": "Slack",
												"begin_at": "2022-01-24T00:00:00.000Z",
												"end_at": "2022-02-06T00:00:00.000Z",
												"open_at": "09:00:00",
												"close_at": "21:00:00",
												"checkin_at": "09:00:00",
												"checkout_at": "22:00:00",
												"seocho": 300,
												"gaepo": 400,
												"actor": "chaepark",
												"deleted_at": null,
												"updated_at": "2022-01-27T05:37:14.064Z",
												"created_at": "2021-12-31T15:00:00.000Z",
												"_comment": 1
											}
										}
									},
									"Example 2": {
										"value": {
											"status": 200,
											"code": 4040,
											"result": false,
											"message": "[production] 해당 날짜(2022-01-31)의 설정값이 서버에 존재하지 않습니다."
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 400,
											"code": 4000,
											"result": false,
											"message": "업데이트 할 설정 값이 올바르지 않습니다. - production"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 500,
											"code": 5000,
											"result": false,
											"message": "서버 오류입니다.\n잠시 후 재시도하시거나,\n증상이 계속되면 관리자(slack:@ohjongin)에게 문의하세요."
										}
									}
								}
							}
						}
					}
				},
				"description": "클러스터 설정 값 수정",
				"tags": [
					"Config"
				],
				"security": [
					{
						"CookieAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConfigRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "localhost"
		},
		{
			"url": "https://api.checkin.dev.42seoul.io",
			"description": "development"
		}
	]
}